ticket-16.2/paid-connected-user-saloon-no_access rename folder profilcreation_fan_creator_agreements_screen to solve check CI-CD
