+++ lib/src/all_screens.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart' show Widget;
 import 'package:saloonprived_app/src/screens/_dashboard/buy_access_user_dashboard_screen/buy_access_user_dashboard_screen.dart';
 import 'package:saloonprived_app/src/screens/_dashboard/promote_user_dashboard_screen.dart';
 
-import 'package:saloonprived_app/src/screens/profilcreation_fan_creator_agreement_screen/profilcreation_fan_creator_agreement_screen.dart';
-import 'package:saloonprived_app/src/screens/profilcreation_generalcondition_screen/profilcreation_generalcondition_screen.dart';
+import 'package:saloonprived_app/src/screens/profilcreation_fan_creator_agreements_screen/profilcreation_fan_creator_agreements_screen.dart';
+import 'package:saloonprived_app/src/screens/profilcreation_generalconditions_screen/profilcreation_generalconditions_screen.dart';
 
 import 'package:saloonprived_app/src/screens/_activity/current_access_user_subscription_activity_screen/current_access_user_subscription_activity_screen.dart';
 
@@ -42,9 +42,9 @@ import 'screens/notification_screen/notification_screen.dart';
 /// Ce n'est valable QUE pour les SCREENS, les composants n'ont pas besion d'être ici
 final Map<String, Widget Function()> allScreens = {
   'welcome_screen.dart': () => const WelcomeScreen(),
-  'profilcreation_generalcondition_screen.dart': () =>
+  'profilcreation_generalconditions_screen.dart': () =>
       const ProfilcreationGeneralconditionScreen(),
-  'profilcreation_fan_creator_agreement_screen.dart': () =>
+  'profilcreation_fan_creator_agreements_screen.dart': () =>
       const ProfilCreationFanCreatorAgreementScreen(),
   'saloon_for_you_have_subcription_screen.dart': () =>
       const SaloonForYouHaveSubcriptionScreen(),
